Project-FTO/fto-backend/requirements.txt
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
supabase>=2.3.0,<2.5.0
httpx>=0.24.0,<0.26.0
